feat: Add scientific calculator option
Introduces a new setting to enable a scientific calculator mode.

This change adds a `scientificCalculator` boolean to the `Settings` interface and the default settings. It also updates the translation keys to include strings for the scientific calculator.

The `CalculatorScreen.tsx` file has been refactored to use a `StandardCalculator` component, preparing for the introduction of a separate `ScientificCalculator` component in a future commit. The `SettingsModal.tsx` has been updated to include a new 'more' settings page, where the scientific calculator option will eventually reside.